+++ packages/google_maps_flutter/google_maps_flutter/test/fake_maps_controllers.dart
@@ -19,6 +19,7 @@ class FakePlatformGoogleMap {
     updatePolygons(params);
     updatePolylines(params);
     updateCircles(params);
+    updateHeatmaps(params);
     updateTileOverlays(Map.castFrom<dynamic, dynamic, String, dynamic>(params));
   }
 
@@ -84,6 +85,12 @@ class FakePlatformGoogleMap {
 
   Set<Circle> circlesToChange = <Circle>{};
 
+  Set<HeatmapId> heatmapIdsToRemove = <HeatmapId>{};
+
+  Set<Heatmap> heatmapsToAdd = <Heatmap>{};
+
+  Set<Heatmap> heatmapsToChange = <Heatmap>{};
+
   Set<TileOverlayId> tileOverlayIdsToRemove = <TileOverlayId>{};
 
   Set<TileOverlay> tileOverlaysToAdd = <TileOverlay>{};
@@ -110,6 +117,9 @@ class FakePlatformGoogleMap {
         return Future<void>.sync(() {});
       case 'circles#update':
         updateCircles(call.arguments as Map<dynamic, dynamic>?);
+        return Future<void>.sync(() {});
+      case 'heatmaps#update':
+        updateHeatmaps(call.arguments as Map<dynamic, dynamic>?);
         return Future<void>.sync(() {});
       default:
         return Future<void>.sync(() {});
@@ -287,6 +297,16 @@ class FakePlatformGoogleMap {
     circlesToChange = _deserializeCircles(circleUpdates['circlesToChange']);
   }
 
+  void updateHeatmaps(Map<dynamic, dynamic>? heatmapUpdates) {
+    if (heatmapUpdates == null) {
+      return;
+    }
+    heatmapsToAdd = _deserializeHeatmaps(heatmapUpdates['heatmapsToAdd']);
+    heatmapIdsToRemove = _deserializeHeatmapIds(
+        heatmapUpdates['heatmapIdsToRemove'] as List<dynamic>?);
+    heatmapsToChange = _deserializeHeatmaps(heatmapUpdates['heatmapsToChange']);
+  }
+
   void updateTileOverlays(Map<String, dynamic> updateTileOverlayUpdates) {
     if (updateTileOverlayUpdates == null) {
       return;
@@ -338,6 +358,76 @@ class FakePlatformGoogleMap {
       result.add(Circle(
         circleId: CircleId(circleId),
         visible: visible,
+        radius: radius,
+      ));
+    }
+
+    return result;
+  }
+
+  Set<HeatmapId> _deserializeHeatmapIds(List<dynamic>? heatmapIds) {
+    if (heatmapIds == null) {
+      return <HeatmapId>{};
+    }
+    return heatmapIds
+        .map(
+          (dynamic heatmapId) => HeatmapId(heatmapId as String),
+        )
+        .toSet();
+  }
+
+  Set<Heatmap> _deserializeHeatmaps(dynamic heatmaps) {
+    if (heatmaps == null) {
+      return <Heatmap>{};
+    }
+    final List<dynamic> heatmapsData = heatmaps as List<dynamic>;
+    final Set<Heatmap> result = <Heatmap>{};
+    for (final Map<dynamic, dynamic> heatmapData
+        in heatmapsData.cast<Map<dynamic, dynamic>>()) {
+      final String heatmapId = heatmapData['heatmapId'] as String;
+
+      final List<dynamic> dataData = heatmapData['data'] as List<dynamic>;
+      final List<WeightedLatLng> data =
+          dataData.map((dynamic e) => WeightedLatLng.fromJson(e)!).toList();
+
+      final bool dissipating = heatmapData['dissipating'] as bool;
+
+      final Map<String, dynamic>? gradientData =
+          heatmapData['gradient'] as Map<String, dynamic>?;
+      final HeatmapGradient? gradient;
+      if (gradientData != null) {
+        final List<Color> colors = (gradientData['colors'] as List<dynamic>)
+            .cast<int>()
+            .map((int e) => Color(e))
+            .toList();
+        final List<double> startPoints =
+            (gradientData['startPoints'] as List<dynamic>)
+                .cast<double>()
+                .toList();
+
+        final List<HeatmapGradientColor> gradientColors =
+            <HeatmapGradientColor>[];
+        for (int i = 0; i < colors.length; i++) {
+          gradientColors.add(HeatmapGradientColor(colors[i], startPoints[i]));
+        }
+
+        gradient = HeatmapGradient(gradientColors,
+            colorMapSize: gradientData['colorMapSize'] as int);
+      } else {
+        gradient = null;
+      }
+
+      final double? maxIntensity = heatmapData['maxIntensity'] as double?;
+      final double opacity = heatmapData['opacity'] as double;
+      final int radius = heatmapData['radius'] as int;
+
+      result.add(Heatmap(
+        heatmapId: HeatmapId(heatmapId),
+        data: data,
+        dissipating: dissipating,
+        gradient: gradient,
+        maxIntensity: maxIntensity,
+        opacity: opacity,
         radius: radius,
       ));
     }