+++ packages/google_maps_flutter/google_maps_flutter_ios/CHANGELOG.md
@@ -1,5 +1,6 @@
-## NEXT
+## 2.2.0
 
+* Adds support for heatmap layers
 * Updates code for new analysis options.
 * Re-enable XCUITests: testUserInterface.
 * Remove unnecessary `RunnerUITests` target from Podfile of the example app.