+++ packages/google_maps_flutter/google_maps_flutter_platform_interface/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.3.0
 
+* Adds support for heatmap layers.
 * Adds options for gesture handling and tilt controls on web.
 
 ## 2.2.4